etl/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
clickhouse-driver==0.2.9
python-dotenv==1.0.1
apscheduler==3.10.4
